[filter]change filter from float to int type

+++ src/include/mean.h
@@ -11,16 +11,16 @@ extern "C" {
 #include <malloc.h>
 
 struct mean_filter_t {
-	float * buffer;
+	int * buffer;
 	int length;
 	int index;
 	int count;
-	float sum;
+	int sum;
 };
 
 struct mean_filter_t * mean_alloc(int length);
 void mean_free(struct mean_filter_t * filter);
-float mean_update(struct mean_filter_t * filter, float value);
+int mean_update(struct mean_filter_t * filter, int value);
 void mean_clear(struct mean_filter_t * filter);
 
 #ifdef __cplusplus
